+++ src/Resource/Countries.php
@@ -28,7 +28,7 @@ final readonly class Countries extends AbstractResource
         $envelope = $this->transport->get('/countries', $query);
 
         if ($envelope->hasErrors()) {
-            return Result::err($envelope->errors);
+            return Result::err($envelope->errors, $envelope->errorId);
         }
 
         $items = Scalars::toArray($envelope->response);